requirements.txt
pytube
openai-whisper
transformers
torchaudio
yt-dlp
nltk
ffmpeg-python
